+++ Src/FluentAssertions/Primitives/ReferenceTypeAssertions.cs
@@ -15,10 +15,6 @@ namespace FluentAssertions.Primitives
     public abstract class ReferenceTypeAssertions<TSubject, TAssertions>
         where TAssertions : ReferenceTypeAssertions<TSubject, TAssertions>
     {
-        protected ReferenceTypeAssertions()
-        {
-        }
-
         protected ReferenceTypeAssertions(TSubject subject)
         {
             Subject = subject;
@@ -27,7 +23,7 @@ namespace FluentAssertions.Primitives
         /// <summary>
         /// Gets the object which value is being asserted.
         /// </summary>
-        public TSubject Subject { get; protected set; }
+        public TSubject Subject { get; }
 
         /// <summary>
         /// Asserts that the current object has not been initialized yet.