+++ Src/FluentAssertions/Collections/WhichValueConstraint.cs
@@ -11,6 +11,6 @@ namespace FluentAssertions.Collections
         /// <summary>
         /// Gets the value of the object referred to by the key.
         /// </summary>
-        public TValue WhichValue { get; private set; }
+        public TValue WhichValue { get; }
     }
 }