+++ Tests/Shared.Specs/Collections/CollectionAssertionSpecs.cs
@@ -3,7 +3,6 @@ using System.Collections;
 using System.Collections.Generic;
 using System.Diagnostics.CodeAnalysis;
 using System.Linq;
-using FluentAssertions.Collections;
 using Xunit;
 using Xunit.Sdk;
 
@@ -11,16 +10,6 @@ namespace FluentAssertions.Specs
 {
     public class CollectionAssertionSpecs
     {
-        [Fact]
-        public void Should_have_parameterless_constructor()
-        {
-            // Arrange
-            Type type = typeof(CollectionAssertions<,>);
-
-            // Act / Assert
-            type.Should().HaveDefaultConstructor();
-        }
-
         #region Be Null
 
         [Fact]