+++ Tests/Shared.Specs/Types/MethodBaseAssertionSpecs.cs
@@ -1,7 +1,6 @@
 ﻿using System;
 using System.Reflection;
 using FluentAssertions.Common;
-using FluentAssertions.Types;
 using Xunit;
 using Xunit.Sdk;
 
@@ -9,16 +8,6 @@ namespace FluentAssertions.Specs
 {
     public class MethodBaseAssertionSpecs
     {
-        [Fact]
-        public void Should_have_parameterless_constructor()
-        {
-            // Arrange
-            Type type = typeof(MethodBaseAssertions<,>);
-
-            // Act / Assert
-            type.Should().HaveDefaultConstructor();
-        }
-
         #region Return
 
         [Fact]