+++ docs/_pages/releases.md
@@ -14,6 +14,7 @@ sidebar:
 * Add `WithOffset` extension method on `DateTime` for easier creation of `DateTimeOffset` objects.
 
 **Fixes**
+* Reported actual value when it contained `{{{{` or `}}}}` (#1223)
 
 **Breaking Changes**
 * Dropped support for .NET Framework 4.5, .NET Standard 1.3 and 1.6 - [#1227](https://github.com/fluentassertions/fluentassertions/pull/1227)